refactor: decompose createFlowExecutor into smaller modules
Extract logging helpers (saveLog, getRunLog) and run recording
(recordRun) from the 138-line createFlowExecutor closure into
top-level functions, reducing the main function's complexity.

Closes #366
